First working version of migration
