+++ Project.toml
@@ -8,4 +8,5 @@ DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 Jive = "ba5e3d4b-8524-549f-bc71-e76ad9e9deed"
 Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 Serialization = "9e88b42a-f829-5b0c-bbe9-9e923198166b"
+Sockets = "6462fe0b-24de-5631-8697-dd941f90decc"
 ZMQ = "c2297ded-f4af-51ae-bb23-16f91089e4e1"
