+++ test/remotesend/remotesend-base.jl
@@ -1,3 +1,4 @@
+# SPDX-License-Identifier: LGPL-3.0-only
 REMOTE_TEST_PAYLOAD = "Sent remotely"
 MESSAGE_COUNT = 10
